+++ cmd/workspace.go
@@ -22,6 +22,9 @@ THE SOFTWARE.
 package cmd
 
 import (
+	"math"
+	"strings"
+
 	"github.com/hashicorp/go-tfe"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
@@ -137,6 +140,27 @@ func organizationListAll(c TfxClientContext) ([]*tfe.Organization, error) {
 	return allItems, nil
 }
 
+func workspaceListAllRemoteStateConsumers(c TfxClientContext, workspaceId string) ([]*tfe.Workspace, error) {
+	allItems := []*tfe.Workspace{}
+	opts := tfe.RemoteStateConsumersListOptions{
+		ListOptions: tfe.ListOptions{PageNumber: 1, PageSize: 100},
+	}
+	for {
+		items, err := c.Client.Workspaces.ListRemoteStateConsumers(c.Context, workspaceId, &opts)
+		if err != nil {
+			return nil, err
+		}
+
+		allItems = append(allItems, items.Items...)
+		if items.CurrentPage >= items.TotalPages {
+			break
+		}
+		opts.PageNumber = items.NextPage
+	}
+
+	return allItems, nil
+}
+
 func workspaceList(c TfxClientContext, searchString string, runStatus string) error {
 	o.AddMessageUserProvided("List Workspaces for Organization:", c.OrganizationName)
 	items, err := workspaceListAllForOrganization(c, c.OrganizationName, searchString)
@@ -238,12 +262,23 @@ func workspaceShow(c TfxClientContext, workspaceName string) error {
 		logError(err, "failed to read workspace")
 	}
 
+	rsc, err := workspaceListAllRemoteStateConsumers(c, w.ID)
+	if err != nil {
+		return errors.Wrap(err, "failed to list remote state consumers")
+	}
+
+	ta, err := workspaceListAllTeams(c, w.ID, math.MaxInt)
+	if err != nil {
+		return errors.Wrap(err, "failed to list teams")
+	}
+
 	o.AddDeferredMessageRead("ID", w.ID)
 	o.AddDeferredMessageRead("Terraform Version", w.TerraformVersion)
 	o.AddDeferredMessageRead("Execution Mode", w.ExecutionMode)
 	o.AddDeferredMessageRead("Auto Apply", w.AutoApply)
 	o.AddDeferredMessageRead("Working Directory", w.WorkingDirectory)
 	o.AddDeferredMessageRead("Locked", w.Locked)
+	o.AddDeferredMessageRead("Global State Sharing", w.GlobalRemoteState)
 
 	if w.CurrentRun == nil {
 		o.AddDeferredMessageRead("Current Run", "none")
@@ -258,6 +293,30 @@ func workspaceShow(c TfxClientContext, workspaceName string) error {
 		o.AddDeferredMessageRead("Current Run Id", run.ID)
 		o.AddDeferredMessageRead("Current Run Status", run.Status)
 		o.AddDeferredMessageRead("Current Run Created", FormatDateTime(run.CreatedAt))
+	}
+
+	// if there are any Team Assignments,
+	// loop through team access and get team names (requires an additional API call)
+	if len(ta) > 0 {
+		teamNames := []string{}
+		for _, i := range ta {
+			t, err := c.Client.Teams.Read(c.Context, i.Team.ID)
+			if err != nil {
+				return errors.Wrap(err, "failed to find team name")
+			}
+			teamNames = append(teamNames, t.Name)
+		}
+		o.AddDeferredMessageRead("Team Access", strings.Join(teamNames, ","))
+	}
+
+	// if there are any Statefile Sharing with workspaces,
+	// loop through workspace and get names
+	if len(rsc) > 0 {
+		wsNames := []string{}
+		for _, i := range rsc {
+			wsNames = append(wsNames, i.Name)
+		}
+		o.AddDeferredMessageRead("Remote State Sharing", strings.Join(wsNames, ","))
 	}
 
 	return nil