+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 **Added**
 
+* `tfx workspace team list` - List team access information for a given Workspace.
+
 **Changed**
+
+* `tfx workspace show` - Now includes Team Access and Statefile Sharing output for the Workspace.
 
 **Removed**